refactor: better organize and use schemas
