+++ geodjango/samplelocations/api/locations.py
@@ -1,16 +1,9 @@
-from ninja import Router, Schema
+from ninja import Router
 from samplelocations.models import Location
+from samplelocations.schemas import LocationSchema, NotFoundSchema
 from typing import List
 
 router = Router()
-
-class NotFoundSchema(Schema):
-    detail: str
-
-class LocationSchema(Schema):
-    location_id: int
-    coordinates: str
-    date_created: str
 
 @router.get("")
 def get_locations(request) -> List[LocationSchema]:
