+++ geodjango/samplelocations/api/thing.py
@@ -1,5 +1,6 @@
-from ninja import Router, Schema
+from ninja import Router
 from samplelocations.models import Thing, Location, Location_Thing_Junction
+from samplelocations.schemas import FeatureCollection, NotFoundSchema, WellProperties, SpringProperties
 from django.contrib.gis.geos import Point
 from django.shortcuts import get_object_or_404
 from django.forms.models import model_to_dict
@@ -7,53 +8,6 @@ from django.http import HttpResponse
 from typing import List, Tuple
 
 router = Router()
-
-class NotFoundSchema(Schema):
-    detail: str
-
-class GeoJSONGeometry(Schema):
-    """
-    Geometry schema for GeoJSON response.
-    """
-
-    type: str  # e.g., "Point", "LineString", "Polygon"
-    coordinates: (
-        List[float] | List[List[float]] | List[List[List[float]]]
-    )  # Supports Point, LineString, Polygon, etc.
-
-class Feature(Schema):
-    type: str = "Feature"
-    geometry: GeoJSONGeometry
-
-class BaseProperties(Schema):
-    thing_id: int
-    name: str
-    thing_type: str
-    release_status: bool
-    date_created: str
-
-
-class WellProperties(BaseProperties):
-    well_depth_ft: float | None = None
-    hole_depth_ft: float | None = None
-    casing_diameter_ft: float | None = None
-    casing_depth_ft: float | None = None
-    casing_description: str | None = None
-    construction_notes: str | None = None
-
-class SpringProperties(BaseProperties):
-    spring_type: str | None = None
-
-class WellFeature(Feature):
-    properties: WellProperties
-
-class SpringFeature(Feature):
-    properties: SpringProperties
-
-class FeatureCollection(Schema):
-    type: str = "FeatureCollection"
-    features: List = [] # can be WellFeature or SpringFeature. Specifying a union of both types makes the schema include unrelated fields, which is not desired.
-
 
 def get_things(thing_id: int | None = None) -> List[Thing]:
     """
